requirements.txt
beautifulsoup4==4.9.0
click==7.1.2
cmake==3.13.0
dlib==19.7.0
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==1.1.2
html5lib==1.0.1
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.18.4
Pillow==7.1.2
scipy==1.4.1
six==1.14.0
soupsieve==2.0
webencodings==0.5.1
Werkzeug==1.0.1
